refactor(server): split mcp provider responsibilities
